+++ app.js
@@ -1,41 +1,29 @@
-const createError = require("http-errors");
-const express = require("express");
-const path = require("path");
-const cookieParser = require("cookie-parser");
-const logger = require("morgan");
-
-const indexRouter = require("./routes/index");
-const usersRouter = require("./routes/users");
+import {
+  errorHandler,
+  notFoundErrorHandler,
+} from "./app/exceptions/handler.js";
+import router from "./routes/index.js";
+import cookieParser from "cookie-parser";
+import express from "express";
+import logger from "morgan";
 
 const app = express();
 
-// view engine setup
-app.set("views", path.join(__dirname, "views"));
+/**
+ * View engine setup.
+ */
+app.set("views", "views");
 app.set("view engine", "jade");
 
 app.use(logger("dev"));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, "public")));
+app.use(express.static("public"));
 
-app.use("/", indexRouter);
-app.use("/users", usersRouter);
+app.use(router);
 
-// catch 404 and forward to error handler
-app.use(function (req, res, next) {
-  next(createError(404));
-});
+app.use(notFoundErrorHandler);
+app.use(errorHandler);
 
-// error handler
-app.use(function (err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get("env") === "development" ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render("error");
-});
-
-module.exports = app;
+export default app;