+++ routes/users.js
@@ -1,9 +1,6 @@
-const express = require("express");
-const router = express.Router();
+import { Router } from "express";
+const router = Router();
 
-/* GET users listing. */
-router.get("/", function (req, res, next) {
-  res.send("respond with a resource");
-});
+router.get("/", (req, res) => res.json({ data: [] }));
 
-module.exports = router;
+export default router;