+++ tests/TestCase.js
@@ -1,0 +1,6 @@
+import app from "../app.js";
+import supertest from "supertest";
+
+const request = supertest(app);
+
+export { request };