+++ CHANGELOG.md
@@ -2,4 +2,17 @@
 
 All notable changes to `RWID Express.js` will be documented in this file.
 
-## [Unreleased](https://github.com/ngodingbang/rwid-expressjs/compare/v1.0.0...develop)
+## [Unreleased](https://github.com/ngodingbang/rwid-expressjs/compare/1.0.0...develop)
+
+## [1.0.0](https://github.com/ngodingbang/rwid-expressjs/releases/tag/1.0.0) - 2023-12-08
+
+- initial commit by [@ianriizky](https://github.com/ianriizky) in [#2474dd5](https://github.com/ngodingbang/rwid-expressjs/commit/2474dd55a9178543d247825694c8dba150cd6999).
+- feat: add tests script using jest and supertest by [@ianriizky](https://github.com/ianriizky) in [#8a65af8](https://github.com/ngodingbang/rwid-expressjs/commit/8a65af8d0dd483c4863c210ec0902fc3549964f3).
+- feat: run linter using eslint with prettier by [@ianriizky](https://github.com/ianriizky) in [#33f965e](https://github.com/ngodingbang/rwid-expressjs/commit/33f965e27143b9f6dd32acb3ad3e90222d31d459).
+- feat: setup prisma orm by [@ngodingbang](https://github.com/ngodingbang) in [#4846d14](https://github.com/ngodingbang/rwid-expressjs/commit/4846d141ec001995a979fbfe681af72069927fac).
+- chore: update eslint config by [@ngodingbang](https://github.com/ngodingbang) in [#f111ec5](https://github.com/ngodingbang/rwid-expressjs/commit/f111ec587a51a461a282a12e2f4684bb04386279).
+- chore!: move loadEnv function into helpers by [@ngodingbang](https://github.com/ngodingbang) in [#5e96960](https://github.com/ngodingbang/rwid-expressjs/commit/5e969609ba3c263cff4ca5717d2f8f5c234e1a16).
+- feat: create Hash support class for password hashing by [@ngodingbang](https://github.com/ngodingbang) in [#82c332c](https://github.com/ngodingbang/rwid-expressjs/commit/82c332c0856b3829dc82387e1de297711e08574e).
+- chore: create error handler function by [@ngodingbang](https://github.com/ngodingbang) in [#96a20dc](https://github.com/ngodingbang/rwid-expressjs/commit/96a20dcc0a6d794fccd13b598c5530367e3b3b9c).
+- chore: add api documentation page by [@ngodingbang](https://github.com/ngodingbang) in [#be202dd](https://github.com/ngodingbang/rwid-expressjs/commit/be202dd10c122b3994bc8177a0948c681214048c).
+- chore: add error class for NotFound and Unauthorized Error by [@ngodingbang](https://github.com/ngodingbang) in [#418d50a](https://github.com/ngodingbang/rwid-expressjs/commit/418d50a80f5d1bedb0aa419fa5883ec3a2371ac6).